+++ libminifi/test/libtest/unit/TestUtils.h
@@ -167,7 +167,7 @@ class ExceptionSubStringMatcher : public Catch::Matchers::MatcherBase<T> {
 };
 
 bool countLogOccurrencesUntil(const std::string& pattern,
-                              const int occurrences,
+                              const size_t occurrences,
                               const std::chrono::milliseconds max_duration,
                               const std::chrono::milliseconds wait_time = 50ms);
 std::error_code sendMessagesViaTCP(const std::vector<std::string_view>& contents, const asio::ip::tcp::endpoint& remote_endpoint, const std::optional<std::string_view> delimiter = std::nullopt);
